Move NatureCNN to Jax layers
